+++ hw-ledger-transports/browser.ts
@@ -1,20 +1,9 @@
 
-import type { Transport, TransportDef } from './types.ts';
+import LedgerHid from 'https://esm.sh/@ledgerhq/hw-transport-webhid@6.27.12';
+import LedgerUsb from 'https://esm.sh/@ledgerhq/hw-transport-webusb@6.27.12';
 
-import LedgerWebHid from 'https://esm.sh/@ledgerhq/hw-transport-webhid@6.27.12';
-import LedgerWebUsb from 'https://esm.sh/@ledgerhq/hw-transport-webusb@6.27.12';
+import { createDefs } from './util.ts';
 
 export { packageInfo } from './packageInfo.ts';
 
-export const transports: TransportDef[] = [
-  {
-    create: (): Promise<Transport> =>
-      (LedgerWebUsb as unknown as Transport).create(),
-    type: 'webusb'
-  },
-  {
-    create: (): Promise<Transport> =>
-      (LedgerWebHid as unknown as Transport).create(),
-    type: 'hid'
-  }
-];
+export const transports = /*#__PURE__*/ createDefs(['webusb', LedgerUsb], ['hid', LedgerHid]);