+++ hw-ledger-transports/types.ts
@@ -1,13 +1,14 @@
 
-export type LedgerTypes = 'hid' | 'u2f' | 'webusb';
+export type TransportType = 'hid' | 'webusb';
+
 
 export interface Transport {
-  create (): Promise<Transport>;
+  // empty on purpose, just a stub
 }
 
 export interface TransportDef {
   /** Create a transport to be used in Ledger operations */
   create (): Promise<Transport>;
   /** The type of the underlying transport definition */
-  type: LedgerTypes;
+  type: TransportType;
 }