+++ hw-ledger/detectOther.ts
@@ -1,5 +1,5 @@
 
-import { packageInfo as transportInfo } from 'https://deno.land/x/polkadot@0.2.32/hw-ledger-transports/packageInfo.ts';
-import { packageInfo as utilInfo } from 'https://deno.land/x/polkadot@0.2.32/util/packageInfo.ts';
+import { packageInfo as transportInfo } from 'https://deno.land/x/polkadot/hw-ledger-transports/packageInfo.ts';
+import { packageInfo as utilInfo } from 'https://deno.land/x/polkadot/util/packageInfo.ts';
 
 export default [transportInfo, utilInfo];