+++ hw-ledger/types.ts
@@ -1,7 +1,5 @@
 
-import type { HexString } from 'https://deno.land/x/polkadot@0.2.32/util/types.ts';
-
-export type LedgerTypes = 'hid' | 'u2f' | 'webusb';
+import type { HexString } from 'https://deno.land/x/polkadot/util/types.ts';
 
 export interface AccountOptions {
   /** The index of the account */