+++ x-randomvalues/react-native.ts
@@ -2,7 +2,7 @@
 
 import { NativeModules } from 'https://esm.sh/react-native';
 
-import { xglobal } from 'https://deno.land/x/polkadot@0.2.32/x-global/mod.ts';
+import { xglobal } from 'https://deno.land/x/polkadot/x-global/mod.ts';
 
 import { base64Decode } from './base64.ts';
 import { getRandomValues as getRandomValuesGlobal } from './browser.ts';
@@ -19,7 +19,7 @@ interface RNExt {
   }
 }
 
-interface GlobalExt extends Window {
+interface GlobalExt {
   nativeCallSyncHook: unknown;
 }